Use common implementation for getSpeechAfterKey(key)
