+++ tests/system/libraries/ChromeLib.py
@@ -192,17 +192,11 @@ class ChromeLib:
 		"""Ensure speech has stopped, press key, and get speech until it stops.
 		@return: The speech after key press.
 		"""
-		spy = _NvdaLib.getSpyLib()
-		spy.wait_for_speech_to_finish()
-		nextSpeechIndex = spy.get_next_speech_index()
-		spy.emulateKeyPress(key)
-		spy.wait_for_speech_to_finish(speechStartedIndex=nextSpeechIndex)
-		speech = spy.get_speech_at_index_until_now(nextSpeechIndex)
-		return speech
+		return _NvdaLib.getSpeechAfterKey(key)
 
 	@staticmethod
 	def getSpeechAfterTab() -> str:
 		"""Ensure speech has stopped, press tab, and get speech until it stops.
 		@return: The speech after tab.
 		"""
-		return ChromeLib.getSpeechAfterKey('tab')
+		return _NvdaLib.getSpeechAfterKey('tab')
