+++ tests/system/libraries/NotepadLib.py
@@ -121,16 +121,3 @@ class NotepadLib:
 		# Move to the start of file
 		spy.emulateKeyPress('home')
 		spy.wait_for_speech_to_finish()
-
-	@staticmethod
-	def getSpeechAfterKey(key) -> str:
-		"""Ensure speech has stopped, press key, and get speech until it stops.
-		@return: The speech after key press.
-		"""
-		spy = _NvdaLib.getSpyLib()
-		spy.wait_for_speech_to_finish()
-		nextSpeechIndex = spy.get_next_speech_index()
-		spy.emulateKeyPress(key)
-		spy.wait_for_speech_to_finish(speechStartedIndex=nextSpeechIndex)
-		speech = spy.get_speech_at_index_until_now(nextSpeechIndex)
-		return speech
