+++ tests/system/libraries/NvdaLib.py
@@ -392,3 +392,16 @@ def getSpyLib():
 	if spy is None:
 		raise AssertionError("Spy not yet available, check order of keywords and NVDA log for errors.")
 	return spy
+
+
+def getSpeechAfterKey(key) -> str:
+	"""Ensure speech has stopped, press key, and get speech until it stops.
+	@return: The speech after key press.
+	"""
+	spy = getSpyLib()
+	spy.wait_for_speech_to_finish()
+	nextSpeechIndex = spy.get_next_speech_index()
+	spy.emulateKeyPress(key)
+	spy.wait_for_speech_to_finish(speechStartedIndex=nextSpeechIndex)
+	speech = spy.get_speech_at_index_until_now(nextSpeechIndex)
+	return speech
