+++ tests/system/robot/notepadTests.py
@@ -13,6 +13,7 @@ from SystemTestSpy import (
 # Imported for type information
 from NotepadLib import NotepadLib as _NotepadLib
 from AssertsLib import AssertsLib as _AssertsLib
+import NvdaLib as _NvdaLib
 
 _notepad: _NotepadLib = _getLib("NotepadLib")
 _asserts: _AssertsLib = _getLib("AssertsLib")
@@ -41,7 +42,7 @@ def test_symbolLevelWord_all():
 	textStr = ' '.join(_wordsToExpectedSymbolLevelAllSpeech.keys())
 	_notepad.prepareNotepad(f"Test: {textStr}")
 	for expectedWord in _wordsToExpectedSymbolLevelAllSpeech.values():
-		wordSpoken = _notepad.getSpeechAfterKey("numpad6")  # navigate to next word
+		wordSpoken = _NvdaLib.getSpeechAfterKey("numpad6")  # navigate to next word
 		_asserts.strings_match(wordSpoken, expectedWord)
 
 
@@ -49,5 +50,5 @@ def test_symbolLevelWord_default():
 	textStr = ' '.join(_wordsToExpectedSymbolLevelDefaultSpeech.keys())
 	_notepad.prepareNotepad(f"Test: {textStr}")
 	for expectedWord in _wordsToExpectedSymbolLevelDefaultSpeech.values():
-		wordSpoken = _notepad.getSpeechAfterKey("numpad6")  # navigate to next word
+		wordSpoken = _NvdaLib.getSpeechAfterKey("numpad6")  # navigate to next word
 		_asserts.strings_match(wordSpoken, expectedWord)
